requirements.txt
Flask==1.1.2
SQLAlchemy==1.3.18
matplotlib==3.3.1
numpy==1.19.1
MarkupSafe==1.1.1
six==1.15.0
Click==7.1.2
kiwisolver==1.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
cycler==0.10.0
Werkzeug==1.0.1
Jinja2==2.11.2
wheel==0.35.1
Flask-WTF==0.14.3
WTForms==2.3.3
pymssql>=2.1.4
python-dotenv==0.14.0
